+++ lunr/pipeline.py
@@ -143,16 +143,16 @@ class Pipeline:
 
         return tokens
 
-    def run_string(self, string, metadata=None):
+    def run_string(self, string, metadata=None, field_name=None):
         """Convenience method for passing a string through a pipeline and
         getting strings out. This method takes care of wrapping the passed
         string in a token and mapping the resulting tokens back to strings.
 
-        .. note:: This ignores the skipped functions since we can't
-            access field names from this context.
+        .. note:: lunr.js will ignore field names, so do not use `field_name`
+           if you wish to maintain compatibility.
         """
         token = Token(string, metadata)
-        return [str(tkn) for tkn in self.run([token])]
+        return [str(tkn) for tkn in self.run([token], field_name=field_name)]
 
     def reset(self):
         self._stack = []